+++ test/helpers/l10n_setup_widget_tester.dart
@@ -4,6 +4,8 @@ import 'package:hooks_riverpod/hooks_riverpod.dart';
 
 import 'package:projeto_crypto/l10n/core_strings.dart';
 
+import 'widgets/internationalized_texts.dart';
+
 class L10nSetupWidgetTester extends StatelessWidget {
   final Locale locale;
   const L10nSetupWidgetTester({
@@ -34,44 +36,5 @@ class L10nSetupWidgetTester extends StatelessWidget {
         ),
       ),
     ));
-  }
-}
-
-class InternationalizedTexts extends StatelessWidget {
-  const InternationalizedTexts({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Column(
-        children: [
-          Text(CoreString.of(context)!.crypto),
-          Text(CoreString.of(context)!.total),
-          Text(CoreString.of(context)!.port),
-          Text(CoreString.of(context)!.move),
-          Text(CoreString.of(context)!.prices(3)),
-          Text(CoreString.of(context)!.variation(3)),
-          Text(CoreString.of(context)!.quant),
-          Text(CoreString.of(context)!.value),
-          Text(CoreString.of(context)!.conv),
-          Text(CoreString.of(context)!.avai),
-          Text(CoreString.of(context)!.how),
-          Text(CoreString.of(context)!.totEst),
-          Text(CoreString.of(context)!.receive),
-          Text(CoreString.of(context)!.review),
-          Text(CoreString.of(context)!.convert),
-          Text(CoreString.of(context)!.exc),
-          Text(CoreString.of(context)!.comp),
-          Text(CoreString.of(context)!.sucess),
-          Text(CoreString.of(context)!.sucessM),
-          Text(CoreString.of(context)!.writeS),
-          Text(CoreString.of(context)!.theValue),
-          Text(CoreString.of(context)!.youDont),
-          Text(CoreString.of(context)!.zero),
-          Text(CoreString.of(context)!.details),
-          Text(CoreString.of(context)!.pick),
-        ],
-      ),
-    );
   }
 }
